adjust mkdud to work around a problem when updating coreutils (bsc #1087901)
The script that actually applies the DUD uses tools from the coreutils
package (think of 'ln' and 'mv'). For current dists this script has been
adjusted (cee bsc##1087901) but for older dists we have no choice but to
force an update of the script with any DUD that would need it.

Since we have now quite a number of 'fixes' mkdud knows, take the time to
rework the user interface to manage these fixes.