Add bash completion
Handle that mkdud has mandatory options by completing the first
'-', then handle general options --help, --version, and options
selecting a different mode. Only complete directories and *.dud
files in the "show" mode.
In the "create" mode, complete file names and directories. But do
not do that for argument errors or the name. Complete common
distribution names for --dist instead. Assume the default DUD file
name "foo.dud" followed by "--dist" if not enough arguments are
provided. Complete directories and *.iso files instead if the second
create argument is not empty and not an option (e.g. "./"<tab>).
And for --condition, complete "ServicePack0" .. "ServicePack4"
additionally to file names and directories.